blog/core/blog/api/v1/filters.py
from rest_framework import generics
from django_filters import rest_framework as filters
from ...models import Post
from .serializers import PostSerializer


class PostFilter(filters.FilterSet):
    min_price = filters.NumberFilter(field_name="price", lookup_expr="gte")
    max_price = filters.NumberFilter(field_name="price", lookup_expr="lte")

    class Meta:
        model = Post
        fields = [
            "category",
        ]


class PostList(generics.ListAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    filter_backends = (filters.DjangoFilterBackend,)
    filterset_class = PostFilter
